Separated clocklist into multiple files
